% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{aws_get_data}
\alias{aws_get_data}
\title{Download de dados INMET}
\usage{
aws_get_data(id, start, end)
}
\description{
Importa uma série de dados bruta do INMET de uma estação a partir do período
especificado
}
