% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_import.R
\name{aws_import}
\alias{aws_import}
\title{Download de dados INMET}
\usage{
aws_import(id, start, end, proxy = ".")
}
\description{
Importa uma série de dados bruta do INMET de uma estação a partir do período
especificado
}
