% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_station.R
\name{aws_station}
\alias{aws_station}
\title{Get basic information about INMET's automatic weather station}
\usage{
aws_station(only.br = TRUE, proxy = ".")
}
\description{
Get basic information about INMET's automatic weather station
}
