% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{aws_stations}
\alias{aws_stations}
\title{Get basic information about INMET's automatic weather station}
\usage{
aws_stations()
}
\description{
Get basic information about INMET's automatic weather station
}
