% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cws_import.R
\name{cws_import}
\alias{cws_import}
\title{Download de dados INMET}
\usage{
cws_import(id, start = Sys.Date(), end = Sys.Date())
}
\description{
Importa uma série de dados bruta do INMET de uma estação a partir do período
especificado
}
