% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cws_station.R
\name{cws_station}
\alias{cws_station}
\title{Get basic information about INMET's conventional weather station}
\usage{
cws_station(proxy = ".")
}
\description{
Get basic information about INMET's conventional weather station
}
