% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations.R
\name{cws_stations}
\alias{cws_stations}
\title{Get basic information about INMET's conventional weather station}
\usage{
cws_stations()
}
\description{
Get basic information about INMET's conventional weather station
}
