% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{inmet_download}
\alias{inmet_download}
\title{Download de dados INMET
Importa uma série de dados bruta do INMET de uma estação a partir do período
especificado}
\usage{
inmet_download(inicio, fim, estacao)
}
\description{
Download de dados INMET
Importa uma série de dados bruta do INMET de uma estação a partir do período
especificado
}
