% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep.R
\name{inmet_download_bdmep}
\alias{inmet_download_bdmep}
\title{Download de dados de estacoes convencionais no portal BDMEP}
\usage{
inmet_download_bdmep(id, inicio, fim, estacoes = NULL, proxy = ".")
}
\description{
Download de dados de estacoes convencionais no portal BDMEP
}
