% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdmep-convencional.R
\name{inmet_download_bdmep_convencional}
\alias{inmet_download_bdmep_convencional}
\title{Download de dados de estacoes convencionais no portal BDMEP}
\usage{
inmet_download_bdmep_convencional(id, inicio, fim, estacoes = NULL,
  proxy = ".")
}
\description{
Download de dados de estacoes convencionais no portal BDMEP
}
