% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonabra-convencional.R
\name{inmet_download_sonabra_convencional}
\alias{inmet_download_sonabra_convencional}
\title{Download de dados INMET}
\usage{
inmet_download_sonabra_convencional(id, inicio, fim, estacoes = NULL,
  proxy = ".")
}
\description{
Importa uma série de dados bruta do INMET de uma estação a partir do período
especificado
}
