% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations_list.R
\name{aws_stations}
\alias{aws_stations}
\title{Get basic information about INMET's automatic weather station}
\usage{
aws_stations()
}
\description{
This function searches the stations registered in the INMET database.
}
